Porposta Full/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.12
httpx==0.27.0
pydantic==2.9.0
python-docx==1.1.2
openpyxl==3.1.5
aiofiles==23.2.1
